+++ weblate/trans/models/__init__.py
@@ -36,6 +36,7 @@ from weblate.trans.models.component import Component
 from weblate.trans.models.componentlist import AutoComponentList, ComponentList
 from weblate.trans.models.dictionary import Dictionary
 from weblate.trans.models.project import Project
+from weblate.trans.models.shaping import Shaping
 from weblate.trans.models.suggestion import Suggestion, Vote
 from weblate.trans.models.translation import Translation
 from weblate.trans.models.unit import Unit
@@ -59,6 +60,7 @@ __all__ = [
     'WeblateConf',
     'ContributorAgreement',
     'Alert',
+    'Shaping',
 ]