+++ weblate/trans/models/shaping.py
@@ -1,0 +1,36 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright © 2012 - 2020 Michal Čihař <michal@cihar.com>
+#
+# This file is part of Weblate <https://weblate.org/>
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+
+from __future__ import unicode_literals
+
+from django.db import models
+from six import python_2_unicode_compatible
+
+from weblate.trans.fields import RegexField
+
+
+@python_2_unicode_compatible
+class Shaping(models.Model):
+    component = models.ForeignKey("Component", on_delete=models.deletion.CASCADE)
+    shaping_regex = RegexField(max_length=190)
+    key = models.CharField(max_length=190, db_index=True)
+
+    class Meta(object):
+        unique_together = (("key", "component", "shaping_regex"),)