+++ weblate/utils/tests/test_search.py
@@ -200,6 +200,7 @@ class QueryParserTest(TestCase):
         self.assert_query("has:comment", Q(has_comment=True))
         self.assert_query("has:ignored-check", Q(check__ignore=True))
         self.assert_query("has:translation", Q(state__gte=STATE_TRANSLATED))
+        self.assert_query("has:shaping", Q(shaping__isnull=False))
 
     def test_suggestions(self):
         self.assert_query(